requirements.txt
openai==1.59.0
python-dotenv==1.2.1
